// admin-webapp/src/core/types/profile.d.ts
import type { Maybe } from '@/utils/utils';

type PersonData = {
  id: string;
  profileCompleted: Maybe<boolean>;
  salutation: string;
  firstName: string;
  lastName: string;
  dateOfBirth: Maybe<Date>;
  sex: Maybe<string>;
  communicationLanguage: Maybe<string>;
  emailPrivate: string;
  emailBusiness: string;
  phone: string;
  linkedInUrl: string;
  webseiteUrl: string;
  street: string;
  streetNumber: string;
  zip: string;
  city: string;
  entryDate: Maybe<Date>;
  exitDate: Maybe<Date>;
  currentEmployer: string;
  currentJobTitle: string;
  currentSector: string;
  uni: string;
  subject: string;
  uniDegreeYear: string;
  subjectInterests: string;
  dataProtectionTermsAccepted: boolean;
  tags: {
    duties: Maybe<string>;
    helvetialevel: Maybe<string>;
    lastsite: Maybe<string>;
    titledegree: Maybe<string>;
  };
  newsletterSubscribed: boolean;
  registrationComments: string;
};

type ApiData = {
  id: string;
  profileCompleted: Maybe<boolean>;
  salutation: string;
  firstName: string;
  lastName: string;
  dateOfBirth: Maybe<string>;
  sex: Maybe<string>;
  communicationLanguage: Maybe<string>;
  emailPrivate: string;
  emailBusiness: string;
  phone: string;
  linkedInUrl: string;
  webseiteUrl: string;
  street: string;
  streetNumber: string;
  zip: string;
  city: string;
  entryDate: Maybe<string>;
  exitDate: Maybe<string>;
  currentEmployer: string;
  currentJobTitle: string;
  currentSector: string;
  uni: string;
  subject: string;
  uniDegreeYear: string;
  subjectInterests: string;
  dataProtectionTermsAccepted: boolean;
  tags: {
    duties: Maybe<string>;
    helvetialevel: Maybe<string>;
    lastsite: Maybe<string>;
    titledegree: Maybe<string>;
  };
  newsletterSubscribed: boolean;
  registrationComments: string;
};